fix purchase order,invoice,return

+++ app/Http/Controllers/PurchaseOrderController.php
@@ -11,11 +11,7 @@ use App\Http\Requests\UpdatePurchaseOrderRequest;
 
 use App\PurchaseOrder;
 use App\Supplier;
-use App\Product;
-use App\PurchaseReturn;
-use App\PurchaseInvoicePayment;
-use App\BankPurchaseInvoicePayment;
-use DB;
+
 
 class PurchaseOrderController extends Controller
 {
@@ -201,23 +197,29 @@ class PurchaseOrderController extends Controller
         $purchase_order->delete();
 
         //delete all the related data with this purchase order in the database
-        //bank purchase invoice payment related
-        $bank = $purchase_order->purchase_order_invoice->purchase_invoice_payment;
-        foreach ($bank as $key) {
-            \DB::table('bank_purchase_invoice_payment')->where('purchase_invoice_payment_id','=',$key->id)->delete();
+        if(count($purchase_order->purchase_order_invoice)){
+            if($purchase_order->purchase_order_invoice->purchase_invoice_payment->count()){
+                //bank purchase invoice payment related
+                $bank = $purchase_order->purchase_order_invoice->purchase_invoice_payment;
+                foreach ($bank as $key) {
+                    \DB::table('bank_purchase_invoice_payment')->where('purchase_invoice_payment_id','=',$key->id)->delete();
+                }
+                $cash = $purchase_order->purchase_order_invoice->purchase_invoice_payment;
+                foreach ($cash as $key) {
+                    \DB::table('cash_purchase_invoice_payment')->where('purchase_invoice_payment_id','=',$key->id)->delete();
+                }
+            }
         }
         //product related
         \DB::table('product_purchase_order')->where('purchase_order_id','=',$request->purchase_order_id)->delete();
         //invoice related
         \DB::table('purchase_order_invoices')->where('purchase_order_id','=',$request->purchase_order_id)->delete();
-        //purchase invoice payment related
-        \DB::table('purchase_invoice_payments')->where('purchase_order_invoice_id','=',$request->payment_id)->delete();
         //return related
         \DB::table('purchase_returns')->where('purchase_order_id','=',$request->purchase_order_id)->delete();
-
+        //purchase invoice payment related
+        \DB::table('purchase_invoice_payments')->where('purchase_order_invoice_id','=',$request->payment_id)->delete();
         return redirect('purchase-order')
             ->with('successMessage', "Purchase Order has been deleted");
-
     }
 
     public function printPdf(Request $request){
